requirements.txt
langchain-core>=0.1.0
langchain-community>=0.3.0
langchain-ollama>=0.1.0
chromadb
faiss-cpu
transformers
torch
langchain-chroma 